Add test for supersearch
This tests supersearch, but actually tests whether the processor is
processing the same set of crashes in both environments.

This test probably needs to be adjusted since there are race conditions
and other things here, but this is a good start and we can hone it as
we go on.